add post controller
